--- wire/src/budgeting.rs
@@ -1,1 +1,0 @@
-
